Fix lib tests when running the whole suite
This commit fixes a global constant state leaking among the lib
tests. RSpec.describe scope doesn't constrain a `class Foo` definition
to itself, therefore it leaks to the global scope. The simplest
solution to prevent a test to access a class from another frequently
is to wrap the definitions into a module, it doesn't solve the root
cause of the problem because the module is still defined into the
global scope, however, it prevent to happen often.

The root problem was the `TestEvent` class being defined in
`base_event_spec.rb` and `event_dispatcher_spec.rb`.

Reference:
- https://github.com/rspec/rspec-core/issues/2181
- https://makandracards.com/makandra/47189-rspec-how-to-define-classes-for-specs
- https://relishapp.com/rspec/rspec-mocks/docs/mutating-constants