Add open account flow diagram
This commit adds an open account flow diagram using PlantUML,
alongside of task to render `.puml` into `.png` using a docker
container. It's using Use Case Diagram to express the commands, events
and reactors interaction.

It already have some PlantUML skinparam defined as an attempt to keep
consistency of names and colors. As the docs grow, it might be a good
idea to create a base styles file and import it using the builtin
`!include` feature.

It might be also interesting, automate the diagrams rendering process
using GitHub actions.

For create diagrams locally, vscode have a very good extension to
preview changes as you type.

Reference:
- https://plantuml.com/use-case-diagram
- https://plantuml.com/command-line
- https://plantuml.com/skinparam
- https://plantuml.com/stdlib
- https://marketplace.visualstudio.com/items?itemName=jebbs.plantuml