Add flow diagram of a credit transaction happy path
Renames the `base.puml` to `styles.puml` to its intent be more
intuitive when people read it in other files.

Adds a light yellow color constant in the diagram styles.

Adds a custom Dockerfile that copy the style to inside of the
container so we can keep running the command locally using the unix
pipe operator without worry about the file not being found while
rendering diagrams. Alongside, adds a new rake task to create the
container.

Mentions the new rake task in the documentation README.