Add command and event to create a transaction
This commit adds the command "create" and the event "created" for
transaction.

A naming refactor in the transaction model was necessary due to a
conflict with Active Record STI reserved attribute name
"type". Therefore, the transaction "type" was changed to "kind".

The test strategy for the command is quite solid and covers almost
every aspect of the whole integration.
- Ensures a single event was created
- Ensures the event has the metadata attribute properly set
- Ensures a single transaction aggregate was created
- Ensures a transaction starts with PENDING status
- Ensures command and event are forwarding the attributes to
  the aggregate