Add capability to update transaction status
This commits adds a command and event to update the transaction
status. Also, adds a reactor to automatic update the status to
APPROVED after they're created - simulating the happy path flow.

Renames account's CreateValidator reactor to ValidatePending. It was
done to keep consistency on naming between account and transaction
reactors that do the same job.

Removes transaction APPLIED status for now because is possible to
implement a simplified version of the flow which doesn't have
it.

Updates create transaction and open account diagrams.

Refactors some specs to use subject and let.